requirements.txt
Appium-Python-Client==1.0.1
behave==1.2.6
parse==1.15.0
parse-type==0.5.2
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
